implementing PR suggestions
