+++ demo/audio/client/index.js
@@ -17,13 +17,11 @@
 
 import * as tf from '@tensorflow/tfjs';
 import MediaStreamRecorder from 'msr';
-import {ClientAPI, FederatedTfModel, loadAudioTransferLearningModel} from 'federated-learning-client';
+import {ClientAPI} from 'federated-learning-client';
 import {plotSpectrogram, plotSpectrum} from './spectral_plots';
-
-const labelNames = [
-  'one', 'two', 'three', 'four', 'five', 'six', 'seven', 'eight', 'nine',
-  'zero', 'left', 'right', 'go', 'stop'
-];
+import {loadAudioTransferLearningModel, labelNames} from './model';
+import {firstPrompt, nextPrompt} from './prompts';
+import {FrequencyListener} from './audio';
 
 const runOptions = {
   magnitudeThreshold: -40,
@@ -44,22 +42,8 @@ const modelDiv = document.getElementById('model');
 const recordButton = document.getElementById('record-button');
 const modelVersion = document.getElementById('model-version');
 const suggestedLabel = document.getElementById('suggested-label');
-const introText = document.getElementById('intro-text');
 const labeledExamples = document.getElementById('labeled-examples');
-const firstIntro = 'Would you be willing to help me?' +
-    ' I\'d love it if you could show me how to pronounce the word:';
-const laterIntro = 'If you\'re up for another, ' +
-    'could you show me how to pronounce:';
-const basicThanks = ['Thanks!', 'Gracias!', 'Much obliged!', 'Bravo!'];
-let thanksVariants = [];
-thanksVariants = thanksVariants.concat(basicThanks);
-thanksVariants = thanksVariants.concat(basicThanks);
-thanksVariants.push('Not bad!');
-thanksVariants = thanksVariants.concat(basicThanks);
-thanksVariants.push('You\'re getting good at this!');
-thanksVariants = thanksVariants.concat(basicThanks);
-thanksVariants = thanksVariants.concat(basicThanks);
-thanksVariants.push('That was a lot, maybe you should take a break!');
+const introText = document.getElementById('intro-text');
 const waitingTemplate = `Waiting for input&hellip;`;
 const modelTemplate = `
   <div class='chart'>
@@ -101,27 +85,10 @@ function setupUI(stream, model, clientAPI) {
   const randomLabels = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13];
   tf.util.shuffle(randomLabels);
 
-  // Set up audio objects that listen to our stream
-  const AudioContext = window.AudioContext || window.webkitAudioContext;
-  const audioContext = new AudioContext();
-  const source = audioContext.createMediaStreamSource(stream);
-  const analyser = audioContext.createAnalyser();
-  analyser.fftSize = runOptions.frameSize * 2;
-  analyser.smoothingTimeConstant = 0.0;
-  source.connect(analyser);
-  const freqData = new Float32Array(analyser.frequencyBinCount);
-  const rotatingBufferNumFrames =
-      runOptions.numFrames * runOptions.rotatingBufferSizeMultiplier;
-  const rotatingBufferSize =
-      runOptions.modelFFTLength * rotatingBufferNumFrames;
-  const rotatingBuffer = new Float32Array(rotatingBufferSize);
+  // Keep track of counters that change each audio frame or recording iteration
   let resultRow;
   let yTrue;
-
-  // Keep track of counters that change each audio frame or recording iteration
-  let frameCount = 0;
   let labelIdx = 0;
-  let thanksIdx = 0;
   suggestedLabel.innerText = labelNames[randomLabels[labelIdx]];
 
   // Create a recorder to save the raw .wav file
@@ -150,25 +117,16 @@ function setupUI(stream, model, clientAPI) {
 
   // On each audio frame, update our rotating buffer with the latest FFT data
   // from our analyser. For fun, also update the spectrum plot
-  function handleAudioFrame() {
-    analyser.getFloatFrequencyData(freqData);
-    if (freqData[0] === -Infinity && freqData[1] === -Infinity) {
-      return;
-    }
-    const freqDataSlice = freqData.slice(0, runOptions.modelFFTLength);
-    plotSpectrum(spectrumCanvas, freqDataSlice, runOptions);
-    const bufferPos = frameCount % rotatingBufferNumFrames;
-    rotatingBuffer.set(freqDataSlice, bufferPos * runOptions.modelFFTLength);
-    frameCount++;
-  }
-  setInterval(
-      handleAudioFrame,
-      analyser.frequencyBinCount / audioContext.sampleRate * 1000);
+  const listener = FrequencyListener(stream, runOptions);
+  listener.onEachFrame(freqData => {
+    plotSpectrum(spectrumCanvas, freqData, runOptions);
+  });
+  listener.listen();
 
   // Create our record button
-  introText.innerText = firstIntro;
   recordButton.innerHTML = 'Record';
   recordButton.removeAttribute('disabled');
+  introText.innerText = firstPrompt();
 
   // When we click it, record for 1 second
   recordButton.addEventListener('click', () => {
@@ -197,8 +155,7 @@ function setupUI(stream, model, clientAPI) {
     recorder.stop();
 
     // Compute input tensor
-    const freqData = getFrequencyDataFromRotatingBuffer(
-        rotatingBuffer, frameCount - runOptions.numFrames);
+    const freqData = listener.getFrequencyData();
     const x = getInputTensorFromFrequencyData(freqData);
 
     // Plot spectrograms
@@ -255,9 +212,10 @@ function setupUI(stream, model, clientAPI) {
         tf.util.shuffle(randomLabels);  // reshuffle each iteration
       }
       suggestedLabel.innerText = labelNames[randomLabels[labelIdx]];
+
       // thank the user
-      introText.innerText = thanksVariants[thanksIdx] + ' ' + laterIntro;
-      thanksIdx = (thanksIdx + 1) % thanksVariants.length;
+      introText.innerText = nextPrompt();
+
       // re-allow recording
       recordButton.innerText = 'Record';
       recordButton.removeAttribute('disabled');
@@ -273,26 +231,6 @@ function setupUI(stream, model, clientAPI) {
       clientAPI.federatedUpdate(x, y).then(cleanup, cleanup);
     }, cleanup);
   }
-}
-
-function getFrequencyDataFromRotatingBuffer(rotatingBuffer, frameCount) {
-  const size = runOptions.numFrames * runOptions.modelFFTLength;
-  const freqData = new Float32Array(size);
-
-  const rotatingBufferSize = rotatingBuffer.length;
-  const rotatingBufferNumFrames =
-      rotatingBufferSize / runOptions.modelFFTLength;
-  while (frameCount < 0) {
-    frameCount += rotatingBufferNumFrames;
-  }
-  const indexBegin =
-      (frameCount % rotatingBufferNumFrames) * runOptions.modelFFTLength;
-  const indexEnd = indexBegin + size;
-
-  for (let i = indexBegin; i < indexEnd; ++i) {
-    freqData[i - indexBegin] = rotatingBuffer[i % rotatingBufferSize];
-  }
-  return freqData;
 }
 
 function normalize(x) {
