+++ demo/audio/server/package.json
@@ -14,6 +14,7 @@
   },
   "dependencies": {
     "@tensorflow/tfjs": "0.11.7",
+    "@tensorflow/tfjs-node": "^0.1.7",
     "federated-learning-server": "file:../../../src/server"
   },
   "devDependencies": {
