+++ src/client/index.ts
@@ -15,7 +15,5 @@
  * =============================================================================
  */
 
-/** Client code. */
-
 export * from './api';
 export * from './common';
