+++ src/client/package.json
@@ -12,7 +12,7 @@
   },
   "license": "Apache-2.0",
   "scripts": {
-    "publish-local": "rm -rf dist/ && yarn && yarn build && yalc publish",
+    "publish-local": "rm -rf dist/ && yarn && yarn build && yalc push",
     "build": "tsc",
     "lint": "tslint -p . -t verbose"
   },
