+++ test/scripts/build-deps-and-run-tests.sh
@@ -1,4 +1,18 @@
 #!/bin/bash
+# Copyright 2018 Google LLC. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+# =============================================================================
 
 echo '***** locally publish federated-learning-server *****'
 cd ../src/server
@@ -8,10 +22,6 @@ echo '***** locally publish federated-learning-client *****'
 cd ../client
 yarn publish-local
 echo ''
-echo '***** link local libraries *****'
-cd ../../test
-./node_modules/.bin/yalc link federated-learning-client
-./node_modules/.bin/yalc link federated-learning-server
-echo ''
 echo '***** run tests *****'
+cd ../../test
 ts-node node_modules/jasmine/bin/jasmine --config=jasmine.json
